[137] Move freak details to a separate component
